@@ -1,15 +1,12 @@
 name: glade_forms
+resolution: workspace
 description: A universal way to define form validators with support of translations.
 version: 6.0.0
 repository: https://github.com/netglade/glade_forms
 issue_tracker: https://github.com/netglade/glade_forms/issues
-screenshots:
-  - description: The glade_forms package logo.
-    path: doc/icon.png
 
 environment:
-  sdk: ^3.6.0
-resolution: workspace
+  sdk: ^3.10.0
 
 # Add regular dependencies here.
 dependencies:
@@ -24,6 +21,10 @@ dependencies:
   provider: ^6.0.5
 
 dev_dependencies:
-  netglade_analysis: ^18.0.0
+  netglade_analysis: ^21.0.0
   test: ^1.25.8
   yaml: ^3.1.2
+
+screenshots:
+  - description: The glade_forms package logo.
+    path: doc/icon.png
