@@ -2,7 +2,7 @@ name: glade_forms_workspace
 publish_to: "none"
 
 environment:
-  sdk: ">=3.8.0 <4.0.0"
+  sdk: ">=3.10.0"
 
 workspace:
   - glade_forms
