@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:glade_forms_devtools_extension/src/models/glade_model_description.dart';
+import 'package:netglade_flutter_utils/netglade_flutter_utils.dart';
 
-/// Widget to display a single GladeModel instance
+/// Widget to display a single GladeModel instance.
 class ModelCard extends StatelessWidget {
   final GladeModelDescription model;
   final VoidCallback? onTap;
@@ -21,13 +22,13 @@ class ModelCard extends StatelessWidget {
     return Card(
       elevation: isSelected ? 4 : 1,
       color: isSelected ? theme.colorScheme.primaryContainer : null,
-      margin: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
+      margin: const .symmetric(horizontal: 8, vertical: 4),
       child: InkWell(
         onTap: onTap,
         child: Padding(
-          padding: const EdgeInsets.all(12),
+          padding: const .all(12),
           child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
+            crossAxisAlignment: .start,
             children: [
               Row(
                 children: [
@@ -39,14 +40,14 @@ class ModelCard extends StatelessWidget {
                   const SizedBox(width: 8),
                   Expanded(
                     child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
+                      crossAxisAlignment: .start,
                       children: [
                         Text(
                           model.type,
                           style: theme.textTheme.titleSmall?.copyWith(
-                            fontWeight: FontWeight.bold,
+                            fontWeight: .bold,
                           ),
-                          overflow: TextOverflow.ellipsis,
+                          overflow: .ellipsis,
                         ),
                         if (model.isComposed)
                           Text(
@@ -73,48 +74,46 @@ class ModelCard extends StatelessWidget {
                   _StatusChip(
                     label: model.isPure ? 'Pure' : 'Dirty',
                     color: model.isPure ? Colors.blue : Colors.orange,
-                    small: true,
+                    isSmall: true,
                   ),
                   if (model.isUnchanged)
-                    _StatusChip(
+                    const _StatusChip(
                       label: 'Unchanged',
                       color: Colors.grey,
-                      small: true,
+                      isSmall: true,
                     ),
                   if (model.isComposed)
                     _StatusChip(
                       label: '${model.childModels.length} models',
                       color: theme.colorScheme.primary,
-                      small: true,
+                      isSmall: true,
                     )
                   else
                     _StatusChip(
                       label: '${model.inputs.length} inputs',
                       color: Colors.grey,
-                      small: true,
+                      isSmall: true,
                     ),
                 ],
               ),
               if (model.formattedErrors.isNotEmpty) ...[
                 const SizedBox(height: 8),
                 Container(
-                  padding: const EdgeInsets.all(8),
+                  padding: const .all(8),
                   decoration: BoxDecoration(
-                    color: Colors.red.withOpacity(0.1),
-                    borderRadius: BorderRadius.circular(4),
+                    color: Colors.red.withValues(alpha: 0.1),
+                    borderRadius: const .all(.circular(4)),
                   ),
                   child: Row(
+                    spacing: 4,
                     children: [
                       const Icon(Icons.error_outline, size: 16, color: Colors.red),
-                      const SizedBox(width: 4),
                       Expanded(
                         child: Text(
                           model.formattedErrors,
-                          style: theme.textTheme.bodySmall?.copyWith(
-                            color: Colors.red,
-                          ),
+                          style: theme.textTheme.bodySmall?.withColor(Colors.red),
                           maxLines: 2,
-                          overflow: TextOverflow.ellipsis,
+                          overflow: .ellipsis,
                         ),
                       ),
                     ],
@@ -132,32 +131,32 @@ class ModelCard extends StatelessWidget {
 class _StatusChip extends StatelessWidget {
   final String label;
   final Color color;
-  final bool small;
+  final bool isSmall;
 
   const _StatusChip({
     required this.label,
     required this.color,
-    this.small = false,
+    this.isSmall = false,
   });
 
   @override
   Widget build(BuildContext context) {
     return Container(
-      padding: EdgeInsets.symmetric(
-        horizontal: small ? 6 : 8,
-        vertical: small ? 2 : 4,
+      padding: .symmetric(
+        horizontal: isSmall ? 6 : 8,
+        vertical: isSmall ? 2 : 4,
       ),
       decoration: BoxDecoration(
-        color: color.withOpacity(0.2),
-        borderRadius: BorderRadius.circular(12),
-        border: Border.all(color: color.withOpacity(0.5)),
+        color: color.withValues(alpha: 0.2),
+        borderRadius: const .all(.circular(12)),
+        border: .all(color: color.withValues(alpha: 0.5)),
       ),
       child: Text(
         label,
         style: TextStyle(
-          color: color.withOpacity(0.9),
-          fontSize: small ? 11 : 12,
-          fontWeight: FontWeight.w500,
+          color: color.withValues(alpha: 0.9),
+          fontSize: isSmall ? 11 : 12,
+          fontWeight: .w500,
         ),
       ),
     );
