@@ -15,15 +15,14 @@ class StringValidator extends GladeValidator<String> {
     Object? key,
     ShouldValidateCallback<String>? shouldValidate,
     bool allowBlank = true,
-    ValidationSeverity severity = ValidationSeverity.error,
-  }) =>
-      satisfy(
-        (input) => allowBlank ? input.isNotEmpty : input.trim().isNotEmpty,
-        devMessage: devMessage ?? (_) => "Value can't be empty",
-        key: key ?? GladeValidationsKeys.stringEmpty,
-        shouldValidate: shouldValidate,
-        severity: severity,
-      );
+    ValidationSeverity severity = .error,
+  }) => satisfy(
+    (input) => allowBlank ? input.isNotEmpty : input.trim().isNotEmpty,
+    devMessage: devMessage ?? (_) => "Value can't be empty",
+    key: key ?? GladeValidationsKeys.stringEmpty,
+    shouldValidate: shouldValidate,
+    severity: severity,
+  );
 
   /// Checks that value is valid email address.
   ///
@@ -37,23 +36,22 @@ class StringValidator extends GladeValidator<String> {
     bool allowEmpty = false,
     Object? key,
     ShouldValidateCallback<String>? shouldValidate,
-    ValidationSeverity severity = ValidationSeverity.error,
-  }) =>
-      satisfy(
-        (x) {
-          if (x.isEmpty) {
-            return allowEmpty;
-          }
+    ValidationSeverity severity = .error,
+  }) => satisfy(
+    (x) {
+      if (x.isEmpty) {
+        return allowEmpty;
+      }
 
-          final regExp = RegExp(RegexPatterns.email);
+      final regExp = RegExp(RegexPatterns.email);
 
-          return regExp.hasMatch(x);
-        },
-        devMessage: devMessage ?? (value) => 'Value "$value" is not in e-mail format',
-        key: key ?? GladeValidationsKeys.stringNotEmail,
-        shouldValidate: shouldValidate,
-        severity: severity,
-      );
+      return regExp.hasMatch(x);
+    },
+    devMessage: devMessage ?? (value) => 'Value "$value" is not in e-mail format',
+    key: key ?? GladeValidationsKeys.stringNotEmail,
+    shouldValidate: shouldValidate,
+    severity: severity,
+  );
 
   /// Checks that value is valid URL address.
   ///
@@ -66,29 +64,28 @@ class StringValidator extends GladeValidator<String> {
     bool requiresScheme = false,
     Object key = GladeValidationsKeys.stringNotUrl,
     ShouldValidateCallback<String>? shouldValidate,
-    ValidationSeverity severity = ValidationSeverity.error,
-  }) =>
-      satisfy(
-        (value) {
-          if (value.isEmpty) {
-            return allowEmpty;
-          }
+    ValidationSeverity severity = .error,
+  }) => satisfy(
+    (value) {
+      if (value.isEmpty) {
+        return allowEmpty;
+      }
 
-          final quantifier = requiresScheme ? '{1}' : '?';
+      final quantifier = requiresScheme ? '{1}' : '?';
 
-          final exp = RegExp(
-            r'^(?:https?:\/\/)' +
-                quantifier +
-                r'(?:www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b(?:[-a-zA-Z0-9()@:%_\+.~#?&\/=]*)$',
-          );
-
-          return exp.hasMatch(value);
-        },
-        devMessage: devMessage ?? (value) => 'Value "$value" is not valid URL address',
-        key: key,
-        shouldValidate: shouldValidate,
-        severity: severity,
+      final exp = RegExp(
+        r'^(?:https?:\/\/)' +
+            quantifier +
+            r'(?:www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b(?:[-a-zA-Z0-9()@:%_\+.~#?&\/=]*)$',
       );
+
+      return exp.hasMatch(value);
+    },
+    devMessage: devMessage ?? (value) => 'Value "$value" is not valid URL address',
+    key: key,
+    shouldValidate: shouldValidate,
+    severity: severity,
+  );
 
   /// Matches provided regex [pattern].
   ///
@@ -110,20 +107,24 @@ class StringValidator extends GladeValidator<String> {
     OnValidate<String>? devMessage,
     Object? key,
     ShouldValidateCallback<String>? shouldValidate,
-    ValidationSeverity severity = ValidationSeverity.error,
-  }) =>
-      satisfy(
-        (value) {
-          final regex =
-              RegExp(pattern, multiLine: multiline, caseSensitive: caseSensitive, dotAll: dotAll, unicode: unicode);
-
-          return regex.hasMatch(value);
-        },
-        devMessage: devMessage ?? (value) => 'Value "$value" does not match regex',
-        key: key ?? GladeValidationsKeys.stringPatternMatch,
-        shouldValidate: shouldValidate,
-        severity: severity,
+    ValidationSeverity severity = .error,
+  }) => satisfy(
+    (value) {
+      final regex = RegExp(
+        pattern,
+        multiLine: multiline,
+        caseSensitive: caseSensitive,
+        dotAll: dotAll,
+        unicode: unicode,
       );
+
+      return regex.hasMatch(value);
+    },
+    devMessage: devMessage ?? (value) => 'Value "$value" does not match regex',
+    key: key ?? GladeValidationsKeys.stringPatternMatch,
+    shouldValidate: shouldValidate,
+    severity: severity,
+  );
 
   /// String's length has to be greater or equal to provided [length].
   ///
@@ -133,17 +134,16 @@ class StringValidator extends GladeValidator<String> {
     OnValidate<String>? devMessage,
     Object? key,
     ShouldValidateCallback<String>? shouldValidate,
-    ValidationSeverity severity = ValidationSeverity.error,
-  }) =>
-      satisfy(
-        (value) {
-          return value.length >= length;
-        },
-        devMessage: devMessage ?? (value) => 'Value "$value" is shorter than allowed length $length',
-        key: key ?? GladeValidationsKeys.stringMinLength,
-        shouldValidate: shouldValidate,
-        severity: severity,
-      );
+    ValidationSeverity severity = .error,
+  }) => satisfy(
+    (value) {
+      return value.length >= length;
+    },
+    devMessage: devMessage ?? (value) => 'Value "$value" is shorter than allowed length $length',
+    key: key ?? GladeValidationsKeys.stringMinLength,
+    shouldValidate: shouldValidate,
+    severity: severity,
+  );
 
   /// String's length has to be less or equal to provided [length].
   ///
@@ -153,14 +153,13 @@ class StringValidator extends GladeValidator<String> {
     OnValidate<String>? devMessage,
     Object? key,
     ShouldValidateCallback<String>? shouldValidate,
-  }) =>
-      satisfy(
-        (value) => value.length < length,
-        devMessage: devMessage ?? (value) => 'Value "$value " is longer than allowed length $length',
-        key: key ?? GladeValidationsKeys.stringMaxLength,
-        shouldValidate: shouldValidate,
-        metaData: length,
-      );
+  }) => satisfy(
+    (value) => value.length < length,
+    devMessage: devMessage ?? (value) => 'Value "$value " is longer than allowed length $length',
+    key: key ?? GladeValidationsKeys.stringMaxLength,
+    shouldValidate: shouldValidate,
+    metaData: length,
+  );
 
   /// String's length has to be equal to provided [length].
   ///
@@ -170,13 +169,12 @@ class StringValidator extends GladeValidator<String> {
     OnValidate<String>? devMessage,
     Object? key,
     ShouldValidateCallback<String>? shouldValidate,
-    ValidationSeverity severity = ValidationSeverity.error,
-  }) =>
-      satisfy(
-        (value) => value.length == length,
-        devMessage: devMessage ?? (value) => 'Value "$value " has to be $length long characters.',
-        key: key ?? GladeValidationsKeys.stringExactLength,
-        shouldValidate: shouldValidate,
-        severity: severity,
-      );
+    ValidationSeverity severity = .error,
+  }) => satisfy(
+    (value) => value.length == length,
+    devMessage: devMessage ?? (value) => 'Value "$value " has to be $length long characters.',
+    key: key ?? GladeValidationsKeys.stringExactLength,
+    shouldValidate: shouldValidate,
+    severity: severity,
+  );
 }
