@@ -7,17 +7,22 @@ void main() {
   setUp(GladeForms.initialize);
 
   test('GladeInput with non-nullable type', () {
+    // arrange
     final input = GladeInput<int>.create(
       validator: (v) => v.build(),
       value: 0,
       inputKey: 'a',
     );
 
+    // act
+
+    // assert
     expect(input.isValid, isTrue);
   });
 
   group('onChange tests', () {
     test('Setter always trigger change', () {
+      // arrange
       final dependentInput = GladeIntInput(value: -1);
       final input = GladeIntInput(
         value: 0,
@@ -29,13 +34,16 @@ void main() {
       expect(dependentInput.value, equals(-1));
       expect(input.value, equals(0));
 
+      // act
       input.value = 100;
 
+      // assert
       expect(dependentInput.value, equals(100));
       expect(input.value, equals(100));
     });
 
     test('updateValue by default trigger change', () {
+      // arrange
       final dependentInput = GladeIntInput(value: -1);
       final input = GladeIntInput(
         value: 0,
@@ -47,13 +55,16 @@ void main() {
       expect(dependentInput.value, equals(-1));
       expect(input.value, equals(0));
 
+      // act
       input.updateValue(100);
 
+      // assert
       expect(dependentInput.value, equals(100));
       expect(input.value, equals(100));
     });
 
     test('updateValue disables trigger onChange', () {
+      // arrange
       final dependentInput = GladeIntInput(value: -1);
       final input = GladeIntInput(
         value: 0,
@@ -65,8 +76,10 @@ void main() {
       expect(dependentInput.value, equals(-1));
       expect(input.value, equals(0));
 
+      // act
       input.updateValue(100, shouldTriggerOnChange: false);
 
+      // assert
       // ignore: avoid-duplicate-test-assertions, it controlls that onChange wasnt called.
       expect(dependentInput.value, equals(-1));
       expect(input.value, equals(100));
@@ -75,6 +88,7 @@ void main() {
 
   group('onChange with Controller tests', () {
     test('Controller always trigger change', () {
+      // arrange
       final dependentInput = GladeIntInput(value: -1);
       final input = GladeIntInput(
         value: 0,
@@ -87,13 +101,16 @@ void main() {
       expect(dependentInput.value, equals(-1));
       expect(input.value, equals(0));
 
+      // act
       input.controller?.text = '100';
 
+      // assert
       expect(dependentInput.value, equals(100));
       expect(input.value, equals(100));
     });
 
     test('Setter always trigger change', () {
+      // arrange
       final dependentInput = GladeIntInput(value: -1);
       final input = GladeIntInput(
         value: 0,
@@ -106,13 +123,16 @@ void main() {
       expect(dependentInput.value, equals(-1));
       expect(input.value, equals(0));
 
+      // act
       input.value = 100;
 
+      // assert
       expect(dependentInput.value, equals(100));
       expect(input.value, equals(100));
     });
 
     test('updateValue by default trigger change', () {
+      // arrange
       final dependentInput = GladeIntInput(value: -1);
       final input = GladeIntInput(
         value: 0,
@@ -125,13 +145,16 @@ void main() {
       expect(dependentInput.value, equals(-1));
       expect(input.value, equals(0));
 
+      // act
       input.updateValue(100);
 
+      // assert
       expect(dependentInput.value, equals(100));
       expect(input.value, equals(100));
     });
 
     test('updateValue disables trigger onChange', () {
+      // arrange
       final dependentInput = GladeIntInput(value: -1);
       final input = GladeIntInput(
         value: 0,
@@ -144,8 +167,10 @@ void main() {
       expect(dependentInput.value, equals(-1));
       expect(input.value, equals(0));
 
+      // act
       input.updateValue(100, shouldTriggerOnChange: false);
 
+      // assert
       // ignore: avoid-duplicate-test-assertions, it controlls that onChange wasnt called.
       expect(dependentInput.value, equals(-1));
       expect(input.value, equals(100));
@@ -154,7 +179,7 @@ void main() {
 
   group('input with controller triggers onChange correctly', () {
     test('when controller text changes then onChange is called', () {
-      // Arrange
+      // arrange
       final changes = <String>[];
       final input = GladeInput<String>.create(
         value: 'a',
@@ -162,16 +187,16 @@ void main() {
         onChange: (info) => changes.add(info.value),
       );
 
-      // Act
+      // act
       input.controller?.text = 'b';
 
-      // Assert
+      // assert
       expect(input.value, equals('b'));
       expect(changes, equals(['b']));
     });
 
     test('when updateValue is called then onChange is called', () {
-      // Arrange
+      // arrange
       final changes = <String>[];
       final input = GladeInput<String>.create(
         value: 'a',
@@ -179,10 +204,10 @@ void main() {
         onChange: (info) => changes.add(info.value),
       );
 
-      // Act
+      // act
       input.updateValue('b');
 
-      // Assert
+      // assert
       expect(input.value, equals('b'));
       expect(changes, equals(['b']));
     });
@@ -190,7 +215,7 @@ void main() {
     test(
       'when updateValue is called with shouldTriggerOnChange: false then onChange is not called',
       () {
-        // Arrange
+        // arrange
         final changes = <String>[];
         final input = GladeInput<String>.create(
           value: 'a',
@@ -198,10 +223,10 @@ void main() {
           onChange: (info) => changes.add(info.value),
         );
 
-        // Act
+        // act
         input.updateValue('b', shouldTriggerOnChange: false);
 
-        // Assert
+        // assert
         expect(input.value, equals('b'));
         expect(changes, equals([]));
       },
@@ -210,7 +235,7 @@ void main() {
     test(
       'when updateValue is called with shouldTriggerOnChange: false and second updateValue is called without it then only the second call triggers onChange',
       () {
-        // Arrange
+        // arrange
         final changes = <String>[];
         final input = GladeInput<String>.create(
           value: 'a',
@@ -218,13 +243,14 @@ void main() {
           onChange: (info) => changes.add(info.value),
         );
 
-        // Act
+        // act
+
         // Won't trigger onChange.
         input.updateValue('b', shouldTriggerOnChange: false);
         // Should trigger onChange.
         input.updateValue('c');
 
-        // Assert
+        // assert
         expect(input.value, equals('c'));
         expect(changes, equals(['c']));
       },
@@ -233,6 +259,7 @@ void main() {
 
   group('Pure test', () {
     test('ResetToPure', () {
+      // arrange
       final input = GladeIntInput(
         value: 100,
         initialValue: 10,
@@ -242,17 +269,20 @@ void main() {
       expect(input.value, equals(100));
       expect(input.initialValue, equals(10));
 
+      // act
       input.updateValue(0, shouldTriggerOnChange: true);
 
       expect(input.value, equals(0));
 
       input.resetToInitialValue();
 
+      // assert
       expect(input.isPure, isTrue);
       expect(input.value, equals(input.initialValue));
     });
 
     test('SetAsNewPure', () {
+      // arrange
       final input = GladeIntInput(
         initialValue: 20,
         value: 100,
@@ -264,12 +294,14 @@ void main() {
       expect(input.initialValue, equals(20));
       expect(input.isPure, isTrue);
 
+      // act
       // Set as new pure with a new value
       input.setNewInitialValue(
         initialValue: () => 10,
         shouldResetToInitialValue: false,
       );
 
+      // assert
       // Check the new state
       // ignore: avoid-duplicate-test-assertions, check again
       expect(input.value, equals(100));
@@ -281,56 +313,66 @@ void main() {
 
   group('TransformValue', () {
     test('No transformValue sets updated value', () {
-      final input = GladeInput<int>.create(value: 0);
+      // arrange
+      final input = GladeInput.create(value: 0);
 
-      // Act
+      // act
       input.updateValue(5);
 
+      // assert
       expect(input.value, equals(5));
     });
 
     test('No transformValue in nullable type sets updated value to null', () {
+      // arrange
       final input = GladeInput<int?>.create(value: 2);
 
-      // Act
+      // act
       input.updateValue(null);
 
+      // assert
       expect(input.value, isNull);
     });
 
     test('With non-nullable type returns transformed value', () {
+      // arrange
       final input = GladeInput<int>.create(
         value: 0,
         valueTransform: (value) => value * 2,
       );
 
-      // Act
+      // act
       input.updateValue(5);
 
+      // assert
       expect(input.value, equals(10));
     });
 
     test('With nullable type and passed null return null', () {
+      // arrange
       final input = GladeInput<int?>.create(
         value: 2,
         valueTransform: (value) => value == 4 ? null : 20,
       );
 
-      // Act
+      // act
       input.updateValue(4);
 
+      // assert
       expect(input.value, equals(null));
     });
 
     test('With nullable type and passed value returns transformed value', () {
+      // arrange
       final input = GladeInput<int?>.create(
         value: 2,
         valueTransform: (value) => value == 4 ? null : 20,
       );
 
-      // Act
+      // act
       input.updateValue(10);
 
+      // assert
       expect(input.value, equals(20));
     });
   });
