@@ -6,29 +6,38 @@ import 'package:test/test.dart';
 void main() {
   group('notEmpty', () {
     test('When value is empty, notEmpty fails', () {
+      // arrange
       final validator = (StringValidator()..notEmpty()).build();
 
+      // act
       final result = validator.validate('');
 
+      // assert
       expect(result.isValid, isFalse);
       expect(result.errors.firstOrNull?.key, equals(GladeValidationsKeys.stringEmpty));
     });
 
     test('When value is not empty or null, notEmpty pass', () {
+      // arrange
       final validator = (StringValidator()..notEmpty()).build();
 
+      // act
       final result = validator.validate('test x');
 
+      // assert
       expect(result.isValid, isTrue);
     });
   });
 
   group('isEmail', () {
     test('When value is empty, isEmail() fails', () {
+      // arrange
       final validator = (StringValidator()..isEmail()).build();
 
+      // act
       final result = validator.validate('');
 
+      // assert
       expect(result.isValid, isFalse);
       expect(result.errors.firstOrNull?.key, equals(GladeValidationsKeys.stringNotEmail));
     });
@@ -42,10 +51,13 @@ void main() {
       ('a.213.czgmail.com', false),
     ]) {
       test('When email is ${testCase.$1}, isEmail() ${testCase.$2 ? 'pass' : 'fails'}', () {
+        // arrange
         final validator = (StringValidator()..isEmail()).build();
 
+        // act
         final result = validator.validate(testCase.$1);
 
+        // assert
         expect(result.isValid, equals(testCase.$2));
       });
     }
@@ -53,10 +65,13 @@ void main() {
 
   group('isUrl', () {
     test('When value is empty, isUrl() fails', () {
+      // arrange
       final validator = (StringValidator()..isUrl()).build();
 
+      // act
       final result = validator.validate('');
 
+      // assert
       expect(result.isValid, isFalse);
       expect(result.errors.firstOrNull?.key, equals(GladeValidationsKeys.stringNotUrl));
     });
@@ -75,10 +90,13 @@ void main() {
       ('noturl', false, false),
     ]) {
       test('When URL is ${testCase.$1}, isUrl(http: ${testCase.$2}) ${testCase.$3 ? 'pass' : 'fails'}', () {
+        // arrange
         final validator = (StringValidator()..isUrl(requiresScheme: testCase.$2)).build();
 
+        // act
         final result = validator.validate(testCase.$1);
 
+        // assert
         expect(result.isValid, equals(testCase.$3));
       });
     }
@@ -86,27 +104,36 @@ void main() {
 
   group('exactLength()', () {
     test('exactLength() pass', () {
+      // arrange
       final validator = (StringValidator()..exactLength(length: 4)).build();
 
+      // act
       final result = validator.validate('abcd');
 
+      // assert
       expect(result.isValid, isTrue);
     });
 
     test('exactLength(), empty value fails', () {
+      // arrange
       final validator = (StringValidator()..exactLength(length: 4)).build();
 
+      // act
       final result = validator.validate('');
 
+      // assert
       expect(result.isValid, isFalse);
       expect(result.errors.firstOrNull?.key, equals(GladeValidationsKeys.stringExactLength));
     });
 
     test('exactLength() fails', () {
+      // arrange
       final validator = (StringValidator()..exactLength(length: 4)).build();
 
+      // act
       final result = validator.validate('asdasd');
 
+      // assert
       expect(result.isValid, isFalse);
       expect(result.errors.firstOrNull?.key, equals(GladeValidationsKeys.stringExactLength));
     });
@@ -114,26 +141,35 @@ void main() {
 
   group('maxLength()', () {
     test('maxLength() pass', () {
+      // arrange
       final validator = (StringValidator()..maxLength(length: 4)).build();
 
+      // act
       final result = validator.validate('134');
 
+      // assert
       expect(result.isValid, isTrue);
     });
 
     test('maxLength(), empty value pass', () {
+      // arrange
       final validator = (StringValidator()..maxLength(length: 4)).build();
 
+      // act
       final result = validator.validate('');
 
+      // assert
       expect(result.isValid, isTrue);
     });
 
     test('maxLength() fails', () {
+      // arrange
       final validator = (StringValidator()..maxLength(length: 4)).build();
 
+      // act
       final result = validator.validate('asdasd');
 
+      // assert
       expect(result.isValid, isFalse);
       expect(result.errors.firstOrNull?.key, equals(GladeValidationsKeys.stringMaxLength));
     });
@@ -141,27 +177,36 @@ void main() {
 
   group('minLength()', () {
     test('minLength() pass', () {
+      // arrange
       final validator = (StringValidator()..minLength(length: 4)).build();
 
+      // act
       final result = validator.validate('abcd');
 
+      // assert
       expect(result.isValid, isTrue);
     });
 
     test('minLength(), empty value fails', () {
+      // arrange
       final validator = (StringValidator()..minLength(length: 4)).build();
 
+      // act
       final result = validator.validate('');
 
+      // assert
       expect(result.isValid, isFalse);
       expect(result.errors.firstOrNull?.key, equals(GladeValidationsKeys.stringMinLength));
     });
 
     test('minLength() fails', () {
+      // arrange
       final validator = (StringValidator()..minLength(length: 4)).build();
 
+      // act
       final result = validator.validate('a');
 
+      // assert
       expect(result.isValid, isFalse);
       expect(result.errors.firstOrNull?.key, equals(GladeValidationsKeys.stringMinLength));
     });
@@ -169,45 +214,63 @@ void main() {
 
   group('conditional validation', () {
     test('no validator is skipped', () {
-      final validator = (StringValidator()
-            ..minLength(length: 2)
-            ..maxLength(length: 6))
-          .build();
+      // arrange
+      final validator =
+          (StringValidator()
+                ..minLength(length: 2)
+                ..maxLength(length: 6))
+              .build();
 
+      // act
       final result = validator.validate('a');
 
+      // assert
       expect(result.isValid, isFalse);
       expect(
         result.errors.first,
-        isA<ValueSatisfyPredicateError<String>>()
-            .having((x) => x.key, 'Has proper key', equals(GladeValidationsKeys.stringMinLength)),
+        isA<ValueSatisfyPredicateError<String>>().having(
+          (x) => x.key,
+          'Has proper key',
+          equals(GladeValidationsKeys.stringMinLength),
+        ),
       );
     });
 
     test('Min length is skipped', () {
-      final validator = (StringValidator()
-            ..minLength(length: 2, shouldValidate: (_) => false)
-            ..maxLength(length: 6))
-          .build();
+      // arrange
+      final validator =
+          (StringValidator()
+                ..minLength(length: 2, shouldValidate: (_) => false)
+                ..maxLength(length: 6))
+              .build();
 
+      // act
       final result = validator.validate('This string is too long');
 
+      // assert
       expect(result.isValid, isFalse);
       expect(
         result.errors.first,
-        isA<ValueSatisfyPredicateError<String>>()
-            .having((x) => x.key, 'Has proper key', equals(GladeValidationsKeys.stringMaxLength)),
+        isA<ValueSatisfyPredicateError<String>>().having(
+          (x) => x.key,
+          'Has proper key',
+          equals(GladeValidationsKeys.stringMaxLength),
+        ),
       );
     });
 
     test('Max length is skipped', () {
-      final validator = (StringValidator()
-            ..minLength(length: 2)
-            ..maxLength(length: 6, shouldValidate: (_) => false))
-          .build();
+      // arrange
+      final validator =
+          (StringValidator()
+                ..minLength(length: 2)
+                ..maxLength(length: 6, shouldValidate: (_) => false))
+              .build();
 
+      // act
       final result = validator.validate('This string is too long, but it will pass');
 
+      // assert
       expect(result.isValid, isTrue);
     });
   });
