@@ -14,9 +14,13 @@ import 'package:glade_forms_devtools_extension/src/widgets/views/loading_view.da
 import 'package:glade_forms_devtools_extension/src/widgets/views/service_unavailable_view.dart';
 import 'package:multi_split_view/multi_split_view.dart';
 
-/// Debug mode flag - set via --dart-define
-const bool _kDebugMode = bool.fromEnvironment('DEBUG_MODE', defaultValue: false);
+/// Debug mode flag - set via --dart-define.
+// ignore: prefer-boolean-prefixes, keep name
+const bool _kDebugMode = bool.fromEnvironment('DEBUG_MODE');
 
+typedef OnChildTapCallback = void Function(GladeModelDescription model, String childId);
+
+// ignore: prefer-match-file-name, keep in same file
 class GladeFormsExtensionScreen extends StatefulWidget {
   const GladeFormsExtensionScreen({super.key});
 
@@ -30,23 +34,27 @@ class _GladeFormsExtensionScreenState extends State<GladeFormsExtensionScreen> {
   GladeModelDescription? _selectedModel;
   String? _selectedChildId; // Track selected child model ID
   bool _isLoading = false;
-  bool _serviceAvailable = false;
+  bool _isServiceAvailable = false;
   Timer? _refreshTimer;
   String? _error;
 
   // Debug mode state
-  MockScenario _currentScenario = MockScenario.composedModel;
-  final bool _debugModeEnabled = _kDebugMode;
+  MockScenario _currentScenario = .composedModel;
+  final bool _isDebugModeEnabled = _kDebugMode;
 
   @override
   void initState() {
     super.initState();
-    if (_debugModeEnabled) {
+    if (_isDebugModeEnabled) {
+      // ignore: avoid-unnecessary-setstate, it is ok
       _loadMockData();
     } else {
+      // ignore: avoid-async-call-in-sync-function, it is ok
       _checkServiceAndLoadData();
-      _refreshTimer = Timer.periodic(GladeFormsConstants.refreshInterval, (_) {
-        _checkServiceAndLoadData(forceRefresh: false);
+      // ignore: prefer-extracting-callbacks, keep inline
+      _refreshTimer = Timer.periodic(Constants.refreshInterval, (_) {
+        // ignore: avoid-async-call-in-sync-function, it is ok
+        _checkServiceAndLoadData(shouldForceRefresh: false);
       });
     }
   }
@@ -57,79 +65,6 @@ class _GladeFormsExtensionScreenState extends State<GladeFormsExtensionScreen> {
     super.dispose();
   }
 
-  Future<void> _checkServiceAndLoadData({bool forceRefresh = true}) async {
-    setState(() {
-      if (forceRefresh) {
-        _isLoading = true;
-        _error = null;
-      }
-    });
-
-    try {
-      final available = await _service.isServiceAvailable();
-
-      setState(() {
-        _serviceAvailable = available;
-      });
-
-      if (available) {
-        await _loadModels();
-      }
-    } catch (e) {
-      setState(() {
-        _error = 'Error checking service: $e';
-        _serviceAvailable = false;
-      });
-    } finally {
-      setState(() {
-        _isLoading = false;
-      });
-    }
-  }
-
-  Future<void> _loadModels() async {
-    try {
-      final models = await _service.fetchModels();
-      if (mounted) {
-        setState(() {
-          _models = models;
-          _error = null;
-          // Update selected model if it's still in the list
-          if (_selectedModel != null) {
-            _selectedModel = models.firstWhere(
-              (m) => m.id == _selectedModel!.id,
-              orElse: () => models.isNotEmpty ? models.first : _selectedModel!,
-            );
-          }
-        });
-      }
-    } catch (e) {
-      if (mounted) {
-        setState(() {
-          _error = 'Error loading models: $e';
-        });
-      }
-    }
-  }
-
-  void _loadMockData() {
-    setState(() {
-      _models = MockDataProvider.getModelsForScenario(_currentScenario);
-      _serviceAvailable = true;
-      _isLoading = false;
-      _error = null;
-      _selectedModel = _models.isNotEmpty ? _models.first : null;
-      _selectedChildId = null;
-    });
-  }
-
-  void _switchMockScenario(MockScenario scenario) {
-    setState(() {
-      _currentScenario = scenario;
-    });
-    _loadMockData();
-  }
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -137,22 +72,22 @@ class _GladeFormsExtensionScreenState extends State<GladeFormsExtensionScreen> {
         title: Row(
           children: [
             const Text('Glade Forms Inspector'),
-            if (_debugModeEnabled) ...[
-              const SizedBox(width: GladeFormsConstants.spacing8),
+            if (_isDebugModeEnabled) ...[
+              const SizedBox(width: Constants.spacing8),
               Container(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: GladeFormsConstants.spacing8,
-                  vertical: GladeFormsConstants.spacing4,
+                padding: const .symmetric(
+                  horizontal: Constants.spacing8,
+                  vertical: Constants.spacing4,
                 ),
-                decoration: BoxDecoration(
+                decoration: const BoxDecoration(
                   color: Colors.orange,
-                  borderRadius: BorderRadius.circular(4),
+                  borderRadius: .all(.circular(4)),
                 ),
                 child: const Text(
-                  GladeFormsConstants.debugBadge,
+                  Constants.debugBadge,
                   style: TextStyle(
-                    fontSize: GladeFormsConstants.debugBadgeFontSize,
-                    fontWeight: FontWeight.bold,
+                    fontSize: Constants.debugBadgeFontSize,
+                    fontWeight: .bold,
                   ),
                 ),
               ),
@@ -160,7 +95,7 @@ class _GladeFormsExtensionScreenState extends State<GladeFormsExtensionScreen> {
           ],
         ),
         actions: [
-          if (_debugModeEnabled) ...[
+          if (_isDebugModeEnabled)
             PopupMenuButton<MockScenario>(
               icon: const Icon(Icons.science),
               tooltip: 'Select Mock Scenario',
@@ -183,30 +118,31 @@ class _GladeFormsExtensionScreenState extends State<GladeFormsExtensionScreen> {
                   child: Text('Multiple Models'),
                 ),
               ],
-            ),
-          ] else ...[
+            )
+          else
             IconButton(
               icon: const Icon(Icons.refresh),
-              onPressed: _isLoading ? null : _checkServiceAndLoadData,
+              onPressed: _isLoading ? null : () => unawaited(_checkServiceAndLoadData()),
               tooltip: 'Refresh',
             ),
-          ],
         ],
       ),
       body: _ExtensionBody(
         isLoading: _isLoading,
-        serviceAvailable: _serviceAvailable,
+        isServiceAvailable: _isServiceAvailable,
         error: _error,
         models: _models,
         selectedModel: _selectedModel,
         selectedChildId: _selectedChildId,
-        onRetry: _checkServiceAndLoadData,
+        onRetry: () => unawaited(_checkServiceAndLoadData()),
+        // ignore: prefer-extracting-callbacks, keep inline
         onModelSelected: (model) {
           setState(() {
             _selectedModel = model;
             _selectedChildId = null;
           });
         },
+        // ignore: prefer-extracting-callbacks, keep inline
         onChildSelected: (model, childId) {
           setState(() {
             _selectedModel = model;
@@ -216,14 +152,103 @@ class _GladeFormsExtensionScreenState extends State<GladeFormsExtensionScreen> {
       ),
     );
   }
+
+  Future<void> _checkServiceAndLoadData({bool shouldForceRefresh = true}) async {
+    if (shouldForceRefresh) {
+      setState(() {
+        _isLoading = true;
+        _error = null;
+      });
+    }
+
+    try {
+      final isAvailable = await _service.isServiceAvailable();
+
+      if (!context.mounted) return;
+
+      setState(() {
+        _isServiceAvailable = isAvailable;
+      });
+
+      if (isAvailable) {
+        await _loadModels();
+      }
+      // ignore: avoid_catches_without_on_clauses, gotta catch them all
+    } catch (e) {
+      setState(() {
+        _error = 'Error checking service: $e';
+        _isServiceAvailable = false;
+      });
+    } finally {
+      setState(() {
+        _isLoading = false;
+      });
+    }
+  }
+
+  Future<void> _loadModels() async {
+    try {
+      final models = await _service.fetchModels();
+      if (mounted) {
+        setState(() {
+          _models = models;
+          _error = null;
+          // Update selected model if it's still in the list
+          if (_selectedModel case final model?) {
+            _selectedModel = models.firstWhere(
+              (m) => m.id == model.id,
+              // ignore: avoid-unsafe-collection-methods, checked by condition
+              orElse: () => models.isNotEmpty ? models.first : model,
+            );
+          }
+        });
+      }
+      // ignore: avoid_catches_without_on_clauses, gotta catch them all
+    } catch (e) {
+      if (mounted) {
+        setState(() {
+          _error = 'Error loading models: $e';
+        });
+      }
+    }
+  }
+
+  void _loadMockData() {
+    setState(() {
+      _models = MockDataProvider.getModelsForScenario(_currentScenario);
+      _isServiceAvailable = true;
+      _isLoading = false;
+      _error = null;
+      _selectedModel = _models.firstOrNull;
+      _selectedChildId = null;
+    });
+  }
+
+  void _switchMockScenario(MockScenario scenario) {
+    setState(() {
+      _currentScenario = scenario;
+    });
+    _loadMockData();
+  }
 }
 
 // Private widgets
 
 class _ExtensionBody extends StatefulWidget {
+  final bool isLoading;
+
+  final bool isServiceAvailable;
+  final String? error;
+  final List<GladeModelDescription> models;
+  final GladeModelDescription? selectedModel;
+  final String? selectedChildId;
+  final VoidCallback onRetry;
+  final ValueChanged<GladeModelDescription> onModelSelected;
+  final OnChildTapCallback onChildSelected;
+
   const _ExtensionBody({
     required this.isLoading,
-    required this.serviceAvailable,
+    required this.isServiceAvailable,
     required this.error,
     required this.models,
     required this.selectedModel,
@@ -233,38 +258,33 @@ class _ExtensionBody extends StatefulWidget {
     required this.onChildSelected,
   });
 
-  final bool isLoading;
-  final bool serviceAvailable;
-  final String? error;
-  final List<GladeModelDescription> models;
-  final GladeModelDescription? selectedModel;
-  final String? selectedChildId;
-  final VoidCallback onRetry;
-  final void Function(GladeModelDescription model) onModelSelected;
-  final void Function(GladeModelDescription model, String childId) onChildSelected;
-
   @override
   State<_ExtensionBody> createState() => _ExtensionBodyState();
 }
 
 class _ExtensionBodyState extends State<_ExtensionBody> {
-  final MultiSplitViewController _controller = MultiSplitViewController(areas: [
-    Area(
-      data: 'sidebar',
-      min: 300,
-      max: 500,
-      size: 350,
-    ),
-    Area(data: 'detail', flex: 1),
-  ]);
+  final MultiSplitViewController _controller = MultiSplitViewController(
+    areas: [
+      // ignore: avoid-undisposed-instances, it is ok
+      Area(data: 'sidebar', min: 300, max: 500, size: 350),
+      // ignore: avoid-undisposed-instances, it is ok
+      Area(data: 'detail', flex: 1),
+    ],
+  );
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
     if (widget.isLoading) {
-      return LoadingView();
+      return const LoadingView();
     }
 
-    if (!widget.serviceAvailable) {
+    if (!widget.isServiceAvailable) {
       return ServiceUnavailableView(onRetry: widget.onRetry);
     }
 
@@ -277,26 +297,25 @@ class _ExtensionBodyState extends State<_ExtensionBody> {
     }
 
     return MultiSplitViewTheme(
-      data: MultiSplitViewThemeData(
-        dividerThickness: 2,
-      ),
+      data: MultiSplitViewThemeData(dividerThickness: 2),
       child: MultiSplitView(
         controller: _controller,
         builder: (context, area) => switch (area.data) {
           'sidebar' => ModelsSidebar(
-              models: widget.models,
-              selectedModel: widget.selectedModel,
-              selectedChildId: widget.selectedChildId,
-              onModelTap: widget.onModelSelected,
-              onChildTap: widget.onChildSelected,
-            ),
+            models: widget.models,
+            selectedModel: widget.selectedModel,
+            selectedChildId: widget.selectedChildId,
+            onModelTap: widget.onModelSelected,
+            onChildTap: widget.onChildSelected,
+          ),
           'detail' => _DetailPane(
-              selectedModel: widget.selectedModel,
-              selectedChildId: widget.selectedChildId,
-            ),
+            selectedModel: widget.selectedModel,
+            selectedChildId: widget.selectedChildId,
+          ),
           _ => const SizedBox.shrink(),
         },
-        sizeUnderflowPolicy: SizeUnderflowPolicy.stretchFirst,
+        sizeUnderflowPolicy: .stretchFirst,
+        // ignore: prefer-extracting-callbacks, keep  inline, prefer-boolean-prefixes
         dividerBuilder: (axis, index, resizable, dragging, highlighted, themeData) {
           return Container(
             color: dragging ? Theme.of(context).primaryColor : Theme.of(context).dividerColor,
@@ -308,30 +327,33 @@ class _ExtensionBodyState extends State<_ExtensionBody> {
 }
 
 class _DetailPane extends StatelessWidget {
+  final GladeModelDescription? selectedModel;
+
+  final String? selectedChildId;
+
   const _DetailPane({
     required this.selectedModel,
     required this.selectedChildId,
   });
 
-  final GladeModelDescription? selectedModel;
-  final String? selectedChildId;
-
   @override
   Widget build(BuildContext context) {
-    if (selectedModel == null) {
+    final model = selectedModel;
+    if (model == null) {
       return Center(
         child: Text(
-          GladeFormsConstants.selectModelMessage,
+          Constants.selectModelMessage,
           style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-                color: Theme.of(context).colorScheme.onSurfaceVariant,
-              ),
+            color: Theme.of(context).colorScheme.onSurfaceVariant,
+          ),
         ),
       );
     }
 
     // If a child model is selected, show its details
-    if (selectedChildId != null && selectedModel!.isComposed) {
-      final childModel = selectedModel!.childModels.firstWhere(
+    if (selectedChildId != null && model.isComposed) {
+      // ignore: avoid-unsafe-collection-methods, checked by condition
+      final childModel = model.childModels.firstWhere(
         (child) => child.id == selectedChildId,
       );
 
@@ -339,11 +361,11 @@ class _DetailPane extends StatelessWidget {
     }
 
     // If the parent composed model is selected, show composed view
-    if (selectedModel!.isComposed) {
-      return ComposedModelView(model: selectedModel!);
+    if (model.isComposed) {
+      return ComposedModelView(model: model);
     }
 
     // For regular models, show detail view
-    return ModelDetailView(model: selectedModel!);
+    return ModelDetailView(model: model);
   }
 }
