@@ -6,27 +6,24 @@ class StateChip extends StatelessWidget {
   final String label;
   final Color color;
 
-  const StateChip({
-    required this.label,
-    required this.color,
-    super.key,
-  });
+  const StateChip({required this.label, required this.color, super.key});
 
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
+
     return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 2),
+      padding: const .symmetric(horizontal: 8, vertical: 2),
       decoration: BoxDecoration(
         color: color.withAlpha(38),
-        borderRadius: BorderRadius.circular(12),
-        border: Border.all(color: color.withAlpha(77), width: 1),
+        borderRadius: const .all(.circular(12)),
+        border: .all(color: color.withAlpha(77)),
       ),
       child: Text(
         label,
         style: theme.textTheme.labelSmall?.copyWith(
           color: color,
-          fontWeight: FontWeight.w600,
+          fontWeight: .w600,
         ),
       ),
     );
