@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
+import 'package:glade_forms_devtools_extension/src/constants.dart';
 import 'package:glade_forms_devtools_extension/src/models/glade_model_description.dart';
 import 'package:glade_forms_devtools_extension/src/widgets/detail/composed/composed_child_model_card.dart';
-import 'package:glade_forms_devtools_extension/src/widgets/detail/composed/state_badge.dart';
+import 'package:glade_forms_devtools_extension/src/widgets/detail/composed/composed_model_state_card.dart';
 
-/// Widget to display a composed model with its child models
+/// Widget to display a composed model with its child models.
 class ComposedModelView extends StatelessWidget {
   final GladeModelDescription model;
 
@@ -14,30 +15,30 @@ class ComposedModelView extends StatelessWidget {
     final theme = Theme.of(context);
 
     return SingleChildScrollView(
-      padding: const EdgeInsets.all(16),
+      padding: const .all(16),
       child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
+        crossAxisAlignment: .start,
         children: [
           // Composed Model header
           Row(
+            spacing: Constants.spacing16,
             children: [
               const Icon(Icons.account_tree, size: 32),
-              const SizedBox(width: 12),
               Expanded(
                 child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
+                  crossAxisAlignment: .start,
                   children: [
                     Text(
                       model.type,
                       style: theme.textTheme.headlineSmall?.copyWith(
-                        fontWeight: FontWeight.bold,
+                        fontWeight: .bold,
                       ),
                     ),
                     Text(
                       'Composed Model',
                       style: theme.textTheme.bodySmall?.copyWith(
                         color: theme.colorScheme.primary,
-                        fontWeight: FontWeight.w500,
+                        fontWeight: .w500,
                       ),
                     ),
                   ],
@@ -60,13 +61,13 @@ class ComposedModelView extends StatelessWidget {
 
           // Child Models section
           Row(
+            spacing: Constants.spacing8,
             children: [
               const Icon(Icons.list, size: 20),
-              const SizedBox(width: 8),
               Text(
                 'Child Models (${model.childModels.length})',
                 style: theme.textTheme.titleMedium?.copyWith(
-                  fontWeight: FontWeight.bold,
+                  fontWeight: .bold,
                 ),
               ),
             ],
@@ -76,16 +77,16 @@ class ComposedModelView extends StatelessWidget {
           if (model.childModels.isEmpty)
             Card(
               child: Padding(
-                padding: const EdgeInsets.all(24),
+                padding: const .all(24),
                 child: Center(
                   child: Column(
+                    spacing: Constants.spacing8,
                     children: [
                       Icon(
                         Icons.inbox_outlined,
                         size: 48,
                         color: theme.colorScheme.onSurfaceVariant.withAlpha((0.5 * 255).toInt()),
                       ),
-                      const SizedBox(height: 8),
                       Text(
                         'No child models',
                         style: theme.textTheme.bodyMedium?.copyWith(
@@ -100,64 +101,6 @@ class ComposedModelView extends StatelessWidget {
           else
             ...model.childModels.map((child) => ComposedChildModelCard(child: child)),
         ],
-      ),
-    );
-  }
-}
-
-class ComposedModelStateCard extends StatelessWidget {
-  final GladeModelDescription model;
-
-  const ComposedModelStateCard({
-    super.key,
-    required this.model,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                const Icon(Icons.info_outline, size: 20),
-                const SizedBox(width: 8),
-                Text(
-                  'Overall State',
-                  style: theme.textTheme.titleMedium?.copyWith(
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            StateBadge(
-              label: 'All Models Valid',
-              value: model.isValid,
-              icon: model.isValid ? Icons.check_circle : Icons.cancel,
-              color: model.validityColor,
-            ),
-            const Divider(height: 16),
-            StateBadge(
-              label: 'All Models Pure',
-              value: model.isPure,
-              icon: model.isPure ? Icons.check_circle : Icons.cancel,
-              color: model.purityColor,
-            ),
-            const Divider(height: 16),
-            StateBadge(
-              label: 'All Models Unchanged',
-              value: model.isUnchanged,
-              icon: model.isUnchanged ? Icons.check : Icons.edit,
-              color: model.changeColor,
-            ),
-          ],
-        ),
       ),
     );
   }
