@@ -1,35 +1,38 @@
 import 'package:flutter/material.dart';
+import 'package:glade_forms_devtools_extension/src/constants.dart';
 import 'package:glade_forms_devtools_extension/src/widgets/common/input_value.dart';
 
 class InfoRow extends StatelessWidget {
   final String label;
+  // ignore: no-object-declaration, value can be of any type
   final Object? value;
-  final bool inverseBoolColors;
+  final bool hasInverseBoolColors;
   final bool hasBackgroundWhitespaceIndicator;
 
   const InfoRow({
-    super.key,
     required this.label,
     required this.value,
-    this.inverseBoolColors = false,
+    super.key,
+    this.hasInverseBoolColors = false,
     this.hasBackgroundWhitespaceIndicator = false,
   });
 
   @override
   Widget build(BuildContext context) {
     return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
+      crossAxisAlignment: .start,
+      spacing: Constants.spacing8,
       children: [
         Text(
           '$label:',
           style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                color: Theme.of(context).colorScheme.onSurfaceVariant,
-              ),
+            color: Theme.of(context).colorScheme.onSurfaceVariant,
+          ),
         ),
         const SizedBox(width: 8),
         InputValue(
           value: value,
-          inverseBoolColors: inverseBoolColors,
+          shouldInverseBoolColors: hasInverseBoolColors,
           hasBackgroundWhitespaceIndicator: hasBackgroundWhitespaceIndicator,
         ),
       ],
