@@ -18,12 +18,12 @@ class _Model extends GladeModel {
     firstName = GladeStringInput(
       initialValue: '',
       validator: (v) => (v..satisfy((value) => value.isNotEmpty)).build(),
-      validationTranslate: (_, __, ___, ____) => firstName.value.isEmpty ? 'First name cannot be empty' : '',
+      validationTranslate: (_, _, _, _) => firstName.value.isEmpty ? 'First name cannot be empty' : '',
     );
     lastName = GladeStringInput(
       initialValue: '',
       validator: (v) => (v..satisfy((value) => value.isNotEmpty)).build(),
-      validationTranslate: (_, __, ___, ____) => lastName.value.isEmpty ? 'Last name cannot be empty' : '',
+      validationTranslate: (_, _, _, _) => lastName.value.isEmpty ? 'Last name cannot be empty' : '',
     );
 
     super.initialize();
@@ -54,7 +54,7 @@ class ComposedExample extends StatelessWidget {
             ),
             GladeFormConsumer<_ComposedModel>(
               builder: (context, model, child) => Padding(
-                padding: const EdgeInsets.only(bottom: 16),
+                padding: const .only(bottom: 16),
                 child: Column(
                   children: [
                     Text(
@@ -94,13 +94,13 @@ class _Form extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 16),
+      padding: const .symmetric(vertical: 8, horizontal: 16),
       child: Card(
         child: Padding(
-          padding: const EdgeInsets.all(8),
+          padding: const .all(8),
           child: Form(
             child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
+              crossAxisAlignment: .start,
               children: [
                 Text('Person #${index + 1}'),
                 Row(
@@ -108,7 +108,7 @@ class _Form extends StatelessWidget {
                   children: [
                     Expanded(
                       child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
+                        crossAxisAlignment: .start,
                         children: [
                           TextFormField(
                             controller: model.firstName.controller,
@@ -124,7 +124,7 @@ class _Form extends StatelessWidget {
                     ),
                     Expanded(
                       child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
+                        crossAxisAlignment: .start,
                         children: [
                           TextFormField(
                             controller: model.lastName.controller,
